fixup! vinculacao

+++ capitulo-5/vinculacao/numeros.rs
@@ -1,8 +1,7 @@
 pub static NUMERO: u32 = 87654321;
 
-static VINTE: u8 = 20;
+static VINTE: u32 = 20;
 
-#[allow(dead_code)]
-pub extern fn retorna_vinte() -> &'static u8 {
+pub extern fn retorna_vinte() -> &'static u32 {
     &VINTE
 }
